+++ src/parser/Parser.spec.ts
@@ -9,7 +9,7 @@ import { PrintStatement, FunctionStatement, NamespaceStatement, ImportStatement 
 import { Range } from 'vscode-languageserver';
 import { DiagnosticMessages } from '../DiagnosticMessages';
 import { isBlock, isCommentStatement, isFunctionStatement, isIfStatement, isIndexedGetExpression } from '../astUtils/reflection';
-import { expectZeroDiagnostics } from '../testHelpers.spec';
+import { expectDiagnostics, expectZeroDiagnostics } from '../testHelpers.spec';
 import { BrsTranspileState } from './BrsTranspileState';
 import { SourceNode } from 'source-map';
 import { BrsFile } from '../files/BrsFile';
@@ -366,6 +366,19 @@ describe('parser', () => {
             `, ParseMode.BrighterScript).diagnostics[0]?.message).not.to.exist;
         });
 
+        it('does not scrap the entire function when encountering unknown parameter type', () => {
+            const parser = parse(`
+                sub test(param1 as unknownType)
+                end sub
+            `);
+            expectDiagnostics(parser, [{
+                ...DiagnosticMessages.functionParameterTypeIsInvalid('param1', 'unknownType')
+            }]);
+            expect(
+                isFunctionStatement(parser.ast.statements[0])
+            ).to.be.true;
+        });
+
         describe('namespace', () => {
             it('allows namespaces declared inside other namespaces', () => {
                 const parser = parse(`
@@ -377,10 +390,9 @@ describe('parser', () => {
                     end namespace
                 `, ParseMode.BrighterScript);
                 expectZeroDiagnostics(parser);
-                // We expect these names to be "as given" in this context, because we aren't
-                // evaluating a full program.
+                // We expect these names to be "as given" in this context, because we aren't evaluating a full program.
                 expect(parser.references.namespaceStatements.map(statement => statement.getName(ParseMode.BrighterScript))).to.deep.equal([
-                    'Level2.Level3',
+                    'Level1.Level2.Level3',
                     'Level1'
                 ]);
             });